Fix all alignment issues with nonzero-depth layer tops and bottoms.
Align element-following/based layer tops with their specified...

... positions at the top of the screen.  TODO: bottoms should align with
BOTTOM of the screen.

Fix adjustedPosition so that the top of a layer (it's position) ...

... lines up with the top of the screen when the top of the screen is t
that position (as opposed to using the center of the screen as the
baseline position).

Layer bottoms now align with the bottom of the page or...

... the layer's position + it's height at the bottom of the screen
regardless of depth.